postgres_to_es/app/requirements.txt
django==3.2
python-dotenv==0.21.0
django-split-settings==1.1.0
django-debug-toolbar==3.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
django-extensions==3.2.1